Correciones en el modulo de Usuario 1

+++ pet/src/main/java/com/pethistory/pet/models/Rol.java
@@ -12,7 +12,7 @@ import lombok.Getter;
 public class Rol {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long idRol;
+    private Integer idRol;
     
     @Column(length = 50)
     private String nomRol;
